// src/zero/schema.gen.ts
const assetsTable = {
  name: "assets",
  columns: {
    id: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
    },
    cusip: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    asset: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    assetName: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
      serverName: "asset_name",
    },
  },
  primaryKey: ["id"],
} as const;
const countersTable = {
  name: "counters",
  columns: {
    id: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    value: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
    },
  },
  primaryKey: ["id"],
} as const;
const cusipQuarterInvestorActivityTable = {
  name: "cusipQuarterInvestorActivity",
  columns: {
    id: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
    },
    cusip: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    ticker: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    quarter: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    numOpen: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
      serverName: "num_open",
    },
    numAdd: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
      serverName: "num_add",
    },
    numReduce: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
      serverName: "num_reduce",
    },
    numClose: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
      serverName: "num_close",
    },
    numHold: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
      serverName: "num_hold",
    },
  },
  primaryKey: ["id"],
  serverName: "cusip_quarter_investor_activity",
} as const;
const entitiesTable = {
  name: "entities",
  columns: {
    id: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    name: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    category: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    description: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    value: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
    },
    createdAt: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
      serverName: "created_at",
    },
  },
  primaryKey: ["id"],
} as const;
const mediumTable = {
  name: "medium",
  columns: {
    id: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    name: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
  },
  primaryKey: ["id"],
} as const;
const messageTable = {
  name: "message",
  columns: {
    id: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    senderId: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
      serverName: "sender_id",
    },
    mediumId: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
      serverName: "medium_id",
    },
    body: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    labels: {
      type: "json",
      optional: true,
      customType: null as unknown as string[],
    },
    timestamp: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
    },
  },
  primaryKey: ["id"],
} as const;
const periodsTable = {
  name: "periods",
  columns: {
    id: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
    },
    period: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
  },
  primaryKey: ["id"],
} as const;
const searchesTable = {
  name: "searches",
  columns: {
    id: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
    },
    cusip: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    code: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    name: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    category: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
  },
  primaryKey: ["id"],
} as const;
const superinvestorsTable = {
  name: "superinvestors",
  columns: {
    id: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
    },
    cik: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    cikName: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
      serverName: "cik_name",
    },
    cikTicker: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
      serverName: "cik_ticker",
    },
    activePeriods: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
      serverName: "active_periods",
    },
  },
  primaryKey: ["id"],
} as const;
const userTable = {
  name: "user",
  columns: {
    id: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    name: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    partner: {
      type: "boolean",
      optional: true,
      customType: null as unknown as boolean,
    },
  },
  primaryKey: ["id"],
} as const;
const userCountersTable = {
  name: "userCounters",
  columns: {
    userId: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
      serverName: "user_id",
    },
    value: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
    },
  },
  primaryKey: ["userId"],
  serverName: "user_counters",
} as const;
const valueQuartersTable = {
  name: "valueQuarters",
  columns: {
    quarter: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    value: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
    },
  },
  primaryKey: ["quarter"],
  serverName: "value_quarters",
} as const;
const mediumRelationships = {
  messages: [
    {
      sourceField: ["id"],
      destField: ["mediumId"],
      destSchema: "message",
      cardinality: "many",
    },
  ],
} as const;
const messageRelationships = {
  sender: [
    {
      sourceField: ["senderId"],
      destField: ["id"],
      destSchema: "user",
      cardinality: "one",
    },
  ],
  medium: [
    {
      sourceField: ["mediumId"],
      destField: ["id"],
      destSchema: "medium",
      cardinality: "one",
    },
  ],
} as const;
const userRelationships = {
  messages: [
    {
      sourceField: ["id"],
      destField: ["senderId"],
      destSchema: "message",
      cardinality: "many",
    },
  ],
} as const;
/**
 * The Zero schema object.
 * This type is auto-generated from your Drizzle schema definition.
 */
export const schema = {
  tables: {
    assets: assetsTable,
    counters: countersTable,
    cusipQuarterInvestorActivity: cusipQuarterInvestorActivityTable,
    entities: entitiesTable,
    medium: mediumTable,
    message: messageTable,
    periods: periodsTable,
    searches: searchesTable,
    superinvestors: superinvestorsTable,
    user: userTable,
    userCounters: userCountersTable,
    valueQuarters: valueQuartersTable,
  },
  relationships: {
    medium: mediumRelationships,
    message: messageRelationships,
    user: userRelationships,
  },
  enableLegacyQueries: false,
  enableLegacyMutators: false,
} as const;

/**
 * Represents the Zero schema type.
 * This type is auto-generated from your Drizzle schema definition.
 */
export type Schema = typeof schema;